requirements.txt
python-telegram-bot
ccxt
requests
numpy
pandas
scipy
matplotlib
plotly
loguru
pydantic
pytest
aiohttp
asyncio
python-dotenv   
websockets  
ta
scikit-learn
pandas
joblib